feat(laton): option to select compiler (#47)
This pull request adds a command-line argument to select an alternative latex compiler to pdflatex from the bash script. This should solve issue #45.
The input is controlled to avoid trying to run a non-supported compiler, and abort the script with an error message if the entered compiler doesn't exist. Documentation is added in the help command, but not in README, like the -d flag.

I haven't detected any weird behavior, but will of course correct it if there's something wrong.
Also, it's my first PR and I'm not sure I followed the right process, please tell me if I did something wrong :D

Regards

Fixes #45.